% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handler.R
\name{fetch_data_from_state}
\alias{fetch_data_from_state}
\title{Fetch arbovirosis data from all cities of a given state in a given timeframe.}
\usage{
fetch_data_from_state(
  state_code,
  ey_start,
  ey_end,
  ew_start = 1,
  ew_end = 52,
  disease = "dengue",
  normalize = FALSE
)
}
\arguments{
\item{state_code}{string The state prefix}

\item{ey_start}{integer The start year.}

\item{ey_end}{integer The end year.}

\item{ew_start}{integer The first epidemiological week of interest. Ranges
from 1 (default) to 52.}

\item{ew_end}{integer The last epidemiological week of interest. Ranges from
1 to 52 (default).}

\item{disease}{string The disease of interest: `chikungunya`,`zika` or
`dengue` (default)}

\item{normalize}{A boolean.
* `TRUE`: the final cases count column will be normalized as
100(n_of_cases)/max(n_of_cases).
* `FALSE` (the default): the final cases count will be just the sum of
cases in the given week.}
}
\value{
A dataframe with the number of cases in the state per epidemiological
  week.
}
\description{
This function uses InfoDengue API (https://info.dengue.mat.br/services/api)
to fetch arbovirosis data from all cities of a given state. It's possible to
get data related to `dengue`, `zika` and `chikungunya`.
}
\examples{
\dontrun{
df <- fetch_data_from_state("SP", 2019, 2023)
}
}
