% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Plot a given list of time series.}
\usage{
plot_time_series(time_series_list, labels, theme = theme_bw())
}
\arguments{
\item{time_series_list}{list List of time series tibbles to be plotted}

\item{labels}{list The list of labels to each time serie. The first time serie
of the `time_series_list` will be labeled by the first label in this
parameter and so on.}

\item{theme}{A theme object from ggplot. Default is theme_bw()}
}
\value{
The plot object.
}
\description{
Given a list of time series and a list of labels, this function will plot the
series and assign a label to each one of them. The first time serie will be
labeled with the first label and so on.
}
\details{
The number of time series in
`time_series_list` is required to be the same as the number of labels in
`labels`. Besides, each time serie is required to have both `x` and `y`
columns, that will be used to plot them.
}
\examples{
\dontrun{
ts1 <- tibble(x = 1:100, y = cumsum(rnorm(100)))
ts2 <- tibble(x = 1:100, y = cumsum(rnorm(100)))
ts3 <- tibble(x = 1:100, y = cumsum(rnorm(100)))

# Create a list of time series tibbles and labels
time_series_tibbles <- list(ts1, ts2, ts3)
labels <- c("Label A", "Label B", "Label C")

plot_time_series(time_series_tibbles, labels)
}
}
